+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 <!-- Please add changes under the Unreleased section that reads 'No current changes' otherwise -->
 
-# Unreleased
+# v4.7.0
 
 - Support element list or instance in `get` for easy extension
 - Configurable project and python in component dispatch code